requirements.txt
altair==5.3.0
altgraph==0.17.4
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cx_Freeze==7.2.0
cycler==0.12.1
distro==1.9.0
dmgbuild==1.6.2
ds-store==1.3.1
et-xmlfile==1.1.0
finnhub-python==2.4.20
flake8==7.1.0
fonttools==4.53.1
frozendict==2.4.4
gitdb==4.0.11
GitPython==3.1.43
html5lib==1.1
idna==3.7
Jinja2==3.1.4
JPype1==1.5.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lxml==5.2.2
mac-alias==2.2.2
macholib==1.16.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
multitasking==0.0.11
numpy==1.26.4
openpyxl==3.1.2
packaging==24.0
pandas==2.2.2
peewee==3.17.6
pillow==10.3.0
platformdirs==4.2.2
protobuf==4.25.3
pyarrow==17.0.0
pycodestyle==2.12.0
pydeck==0.9.1
pyflakes==3.2.0
Pygments==2.18.0
pyinstaller-hooks-contrib==2024.6
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyxirr==0.10.3
referencing==0.35.1
requests==2.31.0
rich==13.7.1
rpds-py==0.18.1
setuptools==70.0.0
six==1.16.0
smmap==5.0.1
soupsieve==2.5
streamlit==1.34.0
tabula-py==2.9.3
tenacity==8.3.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
typing_extensions==4.11.0
tzdata==2024.1
Unidecode==1.3.8
urllib3==2.2.1
webencodings==0.5.1
wheel==0.43.0
xmltodict==0.13.0
yfinance==0.2.41
